refactor magic link template types to be injected with hooks
