+++ admin/templates-tab.php
@@ -39,12 +39,17 @@ class Disciple_Tools_Bulk_Magic_Link_Sender_Tab_Templates {
             'jquery-ui-dialog'
         ], filemtime( dirname( __FILE__ ) . '/js/templates-tab.js' ), true );
 
+        $template_types = apply_filters( 'dt_magic_link_template_types', [
+            'contacts' => [],
+            'default-options' => [],
+        ]);
         wp_localize_script(
             'dt_magic_links_script', 'dt_magic_links', array(
-                'dt_post_types'                => $this->fetch_post_types(),
-                'dt_magic_links_templates'     => Disciple_Tools_Bulk_Magic_Link_Sender_API::fetch_option( Disciple_Tools_Bulk_Magic_Link_Sender_API::$option_dt_magic_links_templates ),
-                'dt_previous_updated_template' => $this->fetch_previous_updated_template(),
-                'dt_languages_icon'            => esc_html( get_template_directory_uri() . '/dt-assets/images/languages.svg' )
+                'dt_post_types'                 => $this->fetch_post_types(),
+                'dt_magic_links_templates'      => Disciple_Tools_Bulk_Magic_Link_Sender_API::fetch_option( Disciple_Tools_Bulk_Magic_Link_Sender_API::$option_dt_magic_links_templates ),
+                'dt_magic_links_template_types' => $template_types,
+                'dt_previous_updated_template'  => $this->fetch_previous_updated_template(),
+                'dt_languages_icon'             => esc_html( get_template_directory_uri() . '/dt-assets/images/languages.svg' )
             )
         );
     }
