+++ composer.json
@@ -3,7 +3,10 @@
         "dealerdirect/phpcodesniffer-composer-installer": "*",
         "wp-coding-standards/wpcs": "3.1.*"
     },
-
+    "scripts": {
+        "lint": "vendor/bin/phpcs",
+        "lintfix": "vendor/bin/phpcbf"
+    },
     "config": {
         "allow-plugins": {
             "dealerdirect/phpcodesniffer-composer-installer": true
